Refactor: 챗봇 파일명과 경로 리팩토링 및 환경변수 설정
