+++ src/main/java/com/example/algoyai/config/WebConfig.java
@@ -15,6 +15,9 @@ import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 @Configuration
 public class WebConfig {
 
+  @Value("${allowed-connection.url}")
+  private String url;
+
   /**
    * CORS 설정을 위한 WebMvcConfigurer 빈을 생성하는 메서드입니다. 이 메서드는 프론트엔드 애플리케이션이 백엔드 API에 대한 크로스 도메인 요청을 허용하도록
    * 설정합니다.
@@ -34,8 +37,8 @@ public class WebConfig {
       public void addCorsMappings(CorsRegistry registry) {
         registry
             .addMapping("/**")
-            .allowedOrigins("http://15.165.235.170", "http://localhost:8081")
-            .allowedMethods("GET", "POST", "PUT", "DELETE", "OPTIONS")
+            .allowedOrigins(url, "http://localhost:8081")
+            .allowedMethods("GET", "DELETE")
             .allowedHeaders("*")
             .allowCredentials(true);
       }
