+++ src/main/java/com/example/algoyai/repository/AllenRepository.java
@@ -1,6 +1,5 @@
 package com.example.algoyai.repository;
 
-import com.example.algoyai.model.entity.ChatMessage;
 import com.example.algoyai.model.entity.QuizRecommend;
 import org.springframework.data.mongodb.repository.MongoRepository;
